add getParentName
